+++ src/main/java/io/github/astrapi69/mystic/crypt/key/SharedSecretEncryptor.java
@@ -40,32 +40,35 @@ public class SharedSecretEncryptor
 
 	SecretKey secretKey;
 	final String provider;
-	final String cipherTransformation;
+	final String cipherAlgorithm;
 
 	IvParameterSpec ivSpec;
 
 	SharedSecretModel model;
-	public SharedSecretEncryptor(final SharedSecretModel model) {
-		this(model.getPrivateKey(), model.getPublicKey(),
-			model.getKeyAgreementAlgorithm(), model.getSecretKeyAlgorithm(), model.getProvider(),
-			model.getCipherTransformation(), model.getIv());
+
+	public SharedSecretEncryptor(final SharedSecretModel model)
+	{
+		this(model.getPrivateKey(), model.getPublicKey(), model.getKeyAgreementAlgorithm(),
+			model.getSecretKeyAlgorithm(), model.getProvider(), model.getCipherAlgorithm(),
+			model.getIv());
 		this.model = model;
 	}
 
 	public SharedSecretEncryptor(final PrivateKey privateKey, final PublicKey publicKey,
 		final String keyAgreementAlgorithm, final String secretKeyAlgorithm, final String provider,
-		final String cipherTransformation, final byte[] iv)
+		final String cipherAlgorithm, final byte[] iv)
 	{
 		ivSpec = new IvParameterSpec(iv);
 		this.provider = provider;
-		this.cipherTransformation = cipherTransformation;
-		secretKey = RuntimeExceptionDecorator.decorate(() -> KeyAgreementFactory
-			.newSharedSecret(privateKey, publicKey, keyAgreementAlgorithm, secretKeyAlgorithm, provider));
+		this.cipherAlgorithm = cipherAlgorithm;
+		secretKey = RuntimeExceptionDecorator
+			.decorate(() -> KeyAgreementFactory.newSharedSecret(privateKey, publicKey,
+				keyAgreementAlgorithm, secretKeyAlgorithm, provider));
 	}
 
 	public byte[] encrypt(final byte[] toEncrypt) throws Exception
 	{
-		Cipher cipher = Cipher.getInstance(cipherTransformation, provider);
+		Cipher cipher = Cipher.getInstance(cipherAlgorithm, provider);
 		cipher.init(Cipher.ENCRYPT_MODE, secretKey, ivSpec);
 		byte[] encryptedCryptData = new byte[cipher.getOutputSize(toEncrypt.length)];
 		int encryptedLength = cipher.update(toEncrypt, 0, toEncrypt.length, encryptedCryptData, 0);