+++ src/test/java/io/github/astrapi69/mystic/crypt/key/SharedSecretEncryptorTest.java
@@ -1,3 +1,27 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2015 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
 package io.github.astrapi69.mystic.crypt.key;
 
 import static org.testng.AssertJUnit.assertEquals;
@@ -66,14 +90,10 @@ public class SharedSecretEncryptorTest
 
 		// with model
 
-		bobSharedSecretModel = SharedSecretModel.builder()
-			.privateKey(keyPairBob.getPrivate())
-			.publicKey(keyPairAlice.getPublic())
-			.keyAgreementAlgorithm("ECDH")
-			.secretKeyAlgorithm("AES")
-			.provider("BC")
-			.cipherTransformation("AES/GCM/NoPadding")
-			.iv(iv).build();
+		bobSharedSecretModel = SharedSecretModel.builder().privateKey(keyPairBob.getPrivate())
+			.publicKey(keyPairAlice.getPublic()).keyAgreementAlgorithm("ECDH")
+			.secretKeyAlgorithm("AES").provider("BC").cipherAlgorithm("AES/GCM/NoPadding").iv(iv)
+			.build();
 
 
 		encryptor = new SharedSecretEncryptor(bobSharedSecretModel);
@@ -83,14 +103,10 @@ public class SharedSecretEncryptorTest
 		encrypted = encryptor.encrypt(expected.getBytes(StandardCharsets.UTF_8));
 		assertNotNull(encrypted);
 
-		aliceSharedSecretModel = SharedSecretModel.builder()
-			.privateKey(keyPairAlice.getPrivate())
-			.publicKey(keyPairBob.getPublic())
-			.keyAgreementAlgorithm("ECDH")
-			.secretKeyAlgorithm("AES")
-			.provider("BC")
-			.cipherTransformation("AES/GCM/NoPadding")
-			.iv(iv).build();
+		aliceSharedSecretModel = SharedSecretModel.builder().privateKey(keyPairAlice.getPrivate())
+			.publicKey(keyPairBob.getPublic()).keyAgreementAlgorithm("ECDH")
+			.secretKeyAlgorithm("AES").provider("BC").cipherAlgorithm("AES/GCM/NoPadding").iv(iv)
+			.build();
 
 		decryptor = new SharedSecretDecryptor(aliceSharedSecretModel);