+++ CHANGELOG.md
@@ -11,11 +11,12 @@ ADDED:
 
 CHANGED:
 
-- update gradle to new version 8.2
+- update gradle to new version 8.2.1
 - update of gradle-plugin dependency 'com.github.ben-manes.versions.gradle.plugin' to new version 0.47.0
 - update of gradle-plugin dependency 'org.ajoberstar.grgit:grgit-gradle' to new version 5.2.0
 - update of gradle-plugin dependency 'com.diffplug.spotless:spotless-plugin-gradle' to new minor version 6.19.0
-- update of crypt-data dependency version to 8.4
+- update of crypt-api dependency version to 8.6
+- update of crypt-data dependency version to 8.5
 - update of dependency commons-codec dependency version to 2.13.0
 - update of dependency commons-io dependency version to 1.16.0
 - update of dependency guava version to new version 32.1.1-jre